mpathsenser: Process and Analyse Data from m-Path Sense

Overcomes one of the major challenges in mobile (passive) sensing, namely being able to pre-process the raw data that comes from a mobile sensing app, specifically "m-Path Sense" https://m-path.io. The main task of 'mpathsenser' is therefore to read "m-Path Sense" JSON files into a database and provide several convenience functions to aid in data processing.

See also

Author

Maintainer: Koen Niemeijer koen.niemeijer@kuleuven.be (ORCID)

Other contributors:

Kristof Meers kristof.meers@kuleuven.be (ORCID) [contributor]

KU Leuven [copyright holder, funder]